Added feature to remember selected GDrive folder
Note: This commit is  a replica of one I made previously, but now it's
in it's own branch. This feature is in response to rojer#18. The feature
allows the user to select a Drive folder from the dropdown and check the
"Remember this folder" box. Next time when the user takes a screenshot,
this folder (and it's children) will be loaded in the dropdown. A
localStorage key was added to store these settings.