Fixes 44 (entire page capture functionality)
This was broken because the extension wasn't properly causing the browser to scroll (`document.body.scroll*` instead of `document.documentElement.scroll*`), so that it could take captures all the way down the page.

The edit page was also incorrectly placing captures tiled along the x-axis b/c it was ignoring the device pixel ratio.